requirements.txt
aiohttp==3.6.2
appier==1.19.16
async-timeout==3.0.1
attrs==19.3.0
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
firebase-admin==4.2.0
Flask==1.1.2
Flask-WTF==0.14.3
gcloud==0.17.0
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
MarkupSafe==1.1.1
msgpack==1.0.0
multidict==4.7.5
oauth2client==3.0.0
Pillow==7.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyee==7.0.1
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2020.1
requests==2.23.0
requests-toolbelt==0.7.0
rsa==4.0
six==1.14.0
slack==0.0.2
slackclient==2.5.0
slackeventsapi==2.1.0
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
yarl==1.4.2
